+++ src/runtime/src/core/io.rs
@@ -1,26 +1,34 @@
+mod buffered_file;
+
+pub use buffered_file::BufferedFile;
+
 use {
-    crate::{runtime_error, ExternalData, ExternalValue, MetaMap, Value, ValueMap},
+    super::string::format,
+    crate::{
+        runtime_error, ExternalData, ExternalValue, KotoFile, KotoRead, KotoWrite, MetaMap, Mutex,
+        RuntimeError, RwLock, Value, ValueMap, Vm,
+    },
     lazy_static::lazy_static,
-    parking_lot::RwLock,
     std::{
         fmt, fs,
-        io::{Read, Seek, SeekFrom, Write},
+        io::{self, BufRead, Read, Seek, SeekFrom, Write},
+        ops::Deref,
         path::{Path, PathBuf},
         sync::Arc,
     },
 };
 
 pub fn make_module() -> ValueMap {
-    use Value::{Bool, Str};
+    use Value::{Bool, Empty, Str};
 
     let mut result = ValueMap::new();
 
     result.add_fn("create", {
         move |vm, args| match vm.get_args(args) {
             [Str(path)] => {
-                let path = Path::new(path.as_str());
+                let path = Path::new(path.as_str()).to_path_buf();
                 match fs::File::create(&path) {
-                    Ok(file) => Ok(File::make_external_value(file, path, false)),
+                    Ok(file) => Ok(File::system_file(file, path)),
                     Err(e) => {
                         return runtime_error!("io.create: Error while creating file: {}", e);
                     }
@@ -34,6 +42,14 @@ pub fn make_module() -> ValueMap {
         }
     });
 
+    result.add_fn("current_dir", |_, _| {
+        let result = match std::env::current_dir() {
+            Ok(path) => Str(path.to_string_lossy().to_string().into()),
+            Err(_) => Empty,
+        };
+        Ok(result)
+    });
+
     result.add_fn("exists", |vm, args| match vm.get_args(args) {
         [Str(path)] => Ok(Bool(Path::new(path.as_str()).exists())),
         _ => runtime_error!("io.exists: Expected path string as argument"),
@@ -42,9 +58,9 @@ pub fn make_module() -> ValueMap {
     result.add_fn("open", {
         move |vm, args| match vm.get_args(args) {
             [Str(path)] => {
-                let path = Path::new(path.as_str());
+                let path = Path::new(path.as_str()).to_path_buf();
                 match fs::File::open(&path) {
-                    Ok(file) => Ok(File::make_external_value(file, path, false)),
+                    Ok(file) => Ok(File::system_file(file, path)),
                     Err(e) => {
                         return runtime_error!("io.open: Error while opening path: {}", e);
                     }
@@ -55,6 +71,28 @@ pub fn make_module() -> ValueMap {
                 unexpected.type_as_string(),
             ),
             _ => runtime_error!("io.open: Expected a String as argument"),
+        }
+    });
+
+    result.add_fn("print", |vm, args| {
+        let result = match vm.get_args(args) {
+            [Str(s)] => vm.stdout().write_line(s.as_str()),
+            [value] => vm.stdout().write_line(&value.to_string()),
+            [Str(format), format_args @ ..] => {
+                let format = format.clone();
+                let format_args = format_args.to_vec();
+                let vm = vm.child_vm();
+                match format::format_string(vm, &format, &format_args) {
+                    Ok(result) => vm.stdout().write_line(&result),
+                    Err(error) => Err(error),
+                }
+            }
+            _ => return runtime_error!("io.print: Expected a string as format argument"),
+        };
+
+        match result {
+            Ok(_) => Ok(Empty),
+            Err(e) => Err(e.with_prefix("string.print")),
         }
     });
 
@@ -87,6 +125,10 @@ pub fn make_module() -> ValueMap {
         }
     });
 
+    result.add_fn("stderr", |vm, _| Ok(File::stderr(vm)));
+    result.add_fn("stdin", |vm, _| Ok(File::stdin(vm)));
+    result.add_fn("stdout", |vm, _| Ok(File::stdout(vm)));
+
     result.add_fn("temp_dir", {
         |_, _| Ok(Str(std::env::temp_dir().to_string_lossy().as_ref().into()))
     });
@@ -95,29 +137,33 @@ pub fn make_module() -> ValueMap {
 }
 
 lazy_static! {
-    static ref FILE_META: Arc<RwLock<MetaMap>> = {
-        use Value::{Number, Str};
+    pub static ref FILE_META: Arc<RwLock<MetaMap>> = {
+        use Value::{Empty, Number, Str};
 
         let mut meta = MetaMap::with_type_name("File");
 
-        meta.add_named_instance_fn("path", |file: &File, _, _| {
-            Ok(Str(file.path.to_string_lossy().as_ref().into()))
+        meta.add_named_instance_fn_mut("flush", |file: &mut File, _, _| match file.flush() {
+            Ok(_) => Ok(Empty),
+            Err(e) => Err(e.with_prefix("File.flush")),
+        });
+
+        meta.add_named_instance_fn("path", |file: &File, _, _| match file.path() {
+            Ok(path) => Ok(Str(path.into())),
+            Err(e) => Err(e.with_prefix("File.path")),
+        });
+
+        meta.add_named_instance_fn_mut("read_line", |file: &mut File, _, _| {
+            match file.read_line() {
+                Ok(Some(result)) => Ok(result[..result.len() - 1].into()),
+                Ok(None) => Ok(Empty),
+                Err(e) => Err(e.with_prefix("File.read_line")),
+            }
         });
 
         meta.add_named_instance_fn_mut("read_to_string", |file: &mut File, _, _| {
-            match file.file.seek(SeekFrom::Start(0)) {
-                Ok(_) => {
-                    let mut buffer = String::new();
-                    match file.file.read_to_string(&mut buffer) {
-                        Ok(_) => Ok(Str(buffer.into())),
-                        Err(e) => {
-                            runtime_error!("File.read_to_string: Error while reading data: {}", e,)
-                        }
-                    }
-                }
-                Err(e) => {
-                    runtime_error!("File.read_to_string: Error while seeking in file: {}", e)
-                }
+            match file.read_to_string() {
+                Ok(result) => Ok(result.into()),
+                Err(e) => Err(e.with_prefix("File.read_to_string")),
             }
         });
 
@@ -126,11 +172,9 @@ lazy_static! {
                 if *n < 0.0 {
                     return runtime_error!("File.seek: Negative seek positions not allowed");
                 }
-                match file.file.seek(SeekFrom::Start(n.into())) {
+                match file.seek(n.into()) {
                     Ok(_) => Ok(Value::Empty),
-                    Err(e) => {
-                        runtime_error!("File.seek: Error while seeking in file: {}", e)
-                    }
+                    Err(e) => Err(e.with_prefix("File.seek")),
                 }
             }
             [unexpected] => runtime_error!(
@@ -141,16 +185,10 @@ lazy_static! {
         });
 
         meta.add_named_instance_fn_mut("write", |file: &mut File, _, args| match args {
-            [value] => {
-                let data = format!("{}", value);
-
-                match file.file.write(data.as_bytes()) {
-                    Ok(_) => Ok(Value::Empty),
-                    Err(e) => {
-                        runtime_error!("File.write: Error while writing to file: {}", e)
-                    }
-                }
-            }
+            [value] => match file.write(value.to_string().as_bytes()) {
+                Ok(_) => Ok(Value::Empty),
+                Err(e) => Err(e.with_prefix("File.write")),
+            },
             _ => runtime_error!("File.write: Expected single value to write as argument"),
         });
 
@@ -162,9 +200,9 @@ lazy_static! {
                     return runtime_error!("File.write_line: Expected single value as argument");
                 }
             };
-            match file.file.write(line.as_bytes()) {
+            match file.write(line.as_bytes()) {
                 Ok(_) => Ok(Value::Empty),
-                Err(e) => runtime_error!("File.write_line: Error while writing to file: {}", e),
+                Err(e) => Err(e.with_prefix("File.write_line")),
             }
         });
 
@@ -172,33 +210,46 @@ lazy_static! {
     };
 }
 
-#[derive(Debug)]
-pub struct File {
-    pub file: fs::File,
-    pub path: PathBuf,
-    /// When set to true the file will be removed when Dropped
-    pub temporary: bool,
+/// The File type used in the io module
+pub struct File(Arc<dyn KotoFile>);
+
+impl Deref for File {
+    type Target = Arc<dyn KotoFile>;
+
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
 }
 
 impl File {
-    pub fn make_external_value(file: fs::File, path: &Path, temporary: bool) -> Value {
+    /// Wraps a file that implements traits typical of a system file in a buffered reader/writer
+    pub fn system_file<T>(file: T, path: PathBuf) -> Value
+    where
+        T: Read + Write + Seek + Send + Sync + 'static,
+    {
         let result = ExternalValue::with_shared_meta_map(
-            File {
-                file,
-                path: path.to_path_buf(),
-                temporary,
-            },
+            File(Arc::new(BufferedSystemFile::new(file, path))),
             FILE_META.clone(),
         );
         Value::ExternalValue(result)
     }
-}
 
-impl Drop for File {
-    fn drop(&mut self) {
-        if self.temporary {
-            let _ = fs::remove_file(&self.path).is_ok();
-        }
+    fn stderr(vm: &Vm) -> Value {
+        let result =
+            ExternalValue::with_shared_meta_map(File(vm.stderr().clone()), FILE_META.clone());
+        Value::ExternalValue(result)
+    }
+
+    fn stdin(vm: &Vm) -> Value {
+        let result =
+            ExternalValue::with_shared_meta_map(File(vm.stdin().clone()), FILE_META.clone());
+        Value::ExternalValue(result)
+    }
+
+    fn stdout(vm: &Vm) -> Value {
+        let result =
+            ExternalValue::with_shared_meta_map(File(vm.stdout().clone()), FILE_META.clone());
+        Value::ExternalValue(result)
     }
 }
 
@@ -210,6 +261,119 @@ impl ExternalData for File {
 
 impl fmt::Display for File {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{}", self.deref())
+    }
+}
+
+impl fmt::Debug for File {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.write_str(&self.to_string())
+    }
+}
+
+struct BufferedSystemFile<T>
+where
+    T: Write,
+{
+    file: Mutex<BufferedFile<T>>,
+    path: PathBuf,
+}
+
+impl<T> BufferedSystemFile<T>
+where
+    T: Read + Write + Seek + Send + Sync,
+{
+    pub fn new(file: T, path: PathBuf) -> Self {
+        Self {
+            file: Mutex::new(BufferedFile::new(file)),
+            path,
+        }
+    }
+}
+
+impl<T> KotoFile for BufferedSystemFile<T>
+where
+    T: Read + Write + Seek + Send + Sync,
+{
+    fn path(&self) -> Result<String, RuntimeError> {
+        Ok(self.path.to_string_lossy().into())
+    }
+
+    fn seek(&self, position: u64) -> Result<(), RuntimeError> {
+        self.file
+            .lock()
+            .seek(SeekFrom::Start(position))
+            .map_err(map_io_err)?;
+        Ok(())
+    }
+}
+
+impl<T> KotoRead for BufferedSystemFile<T>
+where
+    T: Read + Write,
+{
+    fn read_line(&self) -> Result<Option<String>, RuntimeError> {
+        let mut buffer = String::new();
+        match self
+            .file
+            .lock()
+            .read_line(&mut buffer)
+            .map_err(map_io_err)?
+        {
+            0 => Ok(None),
+            _ => Ok(Some(buffer)),
+        }
+    }
+
+    fn read_to_string(&self) -> Result<String, RuntimeError> {
+        let mut buffer = String::new();
+        self.file
+            .lock()
+            .read_to_string(&mut buffer)
+            .map_err(map_io_err)?;
+        Ok(buffer)
+    }
+}
+
+impl<T> KotoWrite for BufferedSystemFile<T>
+where
+    T: Read + Write,
+{
+    fn write(&self, bytes: &[u8]) -> Result<(), RuntimeError> {
+        self.file.lock().write(bytes).map_err(map_io_err)?;
+        Ok(())
+    }
+
+    fn write_line(&self, text: &str) -> Result<(), RuntimeError> {
+        let mut locked = self.file.lock();
+        locked.write(text.as_bytes()).map_err(map_io_err)?;
+        locked.write("\n".as_bytes()).map_err(map_io_err)?;
+        Ok(())
+    }
+
+    fn flush(&self) -> Result<(), RuntimeError> {
+        self.file.lock().flush().map_err(map_io_err)
+    }
+}
+
+impl<T> fmt::Display for BufferedSystemFile<T>
+where
+    T: Write,
+{
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "File({})", self.path.to_string_lossy())
     }
+}
+
+impl<T> fmt::Debug for BufferedSystemFile<T>
+where
+    T: Write,
+{
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.write_str(&self.to_string())
+    }
+}
+
+pub fn map_io_err(e: io::Error) -> RuntimeError {
+    e.to_string().into()
 }