+++ src/runtime/src/lib.rs
@@ -1,14 +1,11 @@
 //! Contains the runtime and core library for the Koto language
 
-pub mod core;
 mod error;
 mod external;
+mod file;
 mod frame;
-mod logger;
 mod meta_map;
-pub mod num2;
-pub mod num4;
-pub mod value;
+mod stdio;
 mod value_iterator;
 mod value_key;
 mod value_list;
@@ -19,16 +16,22 @@ mod value_string;
 mod value_tuple;
 mod vm;
 
+pub mod core;
+pub mod num2;
+pub mod num4;
+pub mod value;
+
 pub use {
     error::*,
     external::{ExternalData, ExternalFunction, ExternalValue},
+    file::{KotoFile, KotoRead, KotoWrite},
     koto_bytecode::{CompilerError, Loader, LoaderError},
     koto_parser::ParserError,
-    logger::{DefaultLogger, KotoLogger},
     meta_map::{BinaryOp, MetaKey, MetaMap, UnaryOp},
     num2::Num2,
     num4::Num4,
-    parking_lot::{RwLock, RwLockReadGuard, RwLockWriteGuard},
+    parking_lot::{Mutex, RwLock, RwLockReadGuard, RwLockWriteGuard},
+    stdio::{DefaultStderr, DefaultStdin, DefaultStdout},
     value::{RuntimeFunction, Value},
     value_iterator::{IntRange, ValueIterator, ValueIteratorOutput},
     value_key::ValueKey,